+++ tests/test_distributed/test_with_dask/test_with_dask.py
@@ -13,7 +13,7 @@ from itertools import starmap
 from math import ceil
 from operator import attrgetter, getitem
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Tuple, Type, Union
+from typing import Any, Dict, Optional, Tuple, Type, Union
 
 import hypothesis
 import numpy as np
@@ -24,6 +24,11 @@ from hypothesis import HealthCheck, given, note, settings
 from sklearn.datasets import make_classification, make_regression
 from xgboost.data import _is_cudf_df
 from xgboost.testing.params import hist_parameter_strategy
+from xgboost.testing.shared import (
+    get_feature_weights,
+    validate_data_initialization,
+    validate_leaf_output,
+)
 
 import xgboost as xgb
 from xgboost import testing as tm
@@ -1223,7 +1228,7 @@ def test_dask_predict_leaf(booster: str, client: "Client") -> None:
     leaf_from_apply = cls.apply(X).reshape(leaf.shape).compute()
     np.testing.assert_allclose(leaf_from_apply, leaf)
 
-    tm.validate_leaf_output(leaf, num_parallel_tree)
+    validate_leaf_output(leaf, num_parallel_tree)
 
 
 def test_dask_iteration_range(client: "Client"):
@@ -1602,7 +1607,7 @@ class TestWithDask:
             fw[i] *= float(i)
         fw = da.from_array(fw)
         parser = os.path.join(tm.demo_dir(__file__), "json-model", "json_parser.py")
-        poly_increasing = tm.get_feature_weights(
+        poly_increasing = get_feature_weights(
             X, y, fw, parser, "approx", model=xgb.dask.DaskXGBRegressor
         )
 
@@ -1610,7 +1615,7 @@ class TestWithDask:
         for i in range(kCols):
             fw[i] *= float(kCols - i)
         fw = da.from_array(fw)
-        poly_decreasing = tm.get_feature_weights(
+        poly_decreasing = get_feature_weights(
             X, y, fw, parser, "approx", model=xgb.dask.DaskXGBRegressor
         )
 
@@ -1710,7 +1715,7 @@ class TestWithDask:
         from sklearn.datasets import load_digits
         X, y = load_digits(return_X_y=True)
         X, y = dd.from_array(X, chunksize=32), dd.from_array(y, chunksize=32)
-        tm.validate_data_initialization(
+        validate_data_initialization(
             xgb.dask.DaskDMatrix, xgb.dask.DaskXGBClassifier, X, y
         )
